+++ app/Http/Controllers/ProductsController.php
@@ -5,7 +5,9 @@ namespace App\Http\Controllers;
 use App\Product;
 use Illuminate\Http\Request;
 use App\Stores;
+use App\Favorites;
 use App\SubCategories;
+use App\UserVisitProducts;
 use Illuminate\Support\Str;
 
 
@@ -99,7 +101,7 @@ class ProductsController extends Controller
 
          $product->save();
      
-         return back()->with('success', 'Product deleted successfully');
+         return back()->with('success', 'Product Created successfully');
          
         }
      
@@ -155,12 +157,37 @@ class ProductsController extends Controller
      */
     public function destroy(Product $product)
     {
-        $product->delete();
-    
-        // Recargar la página actual
-        return back()->with('success', 'Product deleted successfully');
-    }
+        // Verificar si el producto está agregado a favoritos
+        if ($this->isProductInFavorites($product->id)) {
+            // Mostrar un mensaje indicando que el producto está agregado a favoritos
+            return back()->with('error', 'Este producto está agregado a favoritos y no se puede eliminar');
+        } else {
+            // Intentar eliminar el producto
+                
+                try {
 
+                    UserVisitProducts::where('user_visit_products.products_id', $product->id)->delete();
+
+                    $product->delete();
+                    
+                } catch (\Throwable $th) {
+                    
+                    return back()->with('error', 'No se puede eliminar este producto porque está asociado a órdenes existentes.');
+
+
+                }
+
+           
+            
+            // Redirigir a alguna otra página o mostrar un mensaje de éxito
+            return back()->with('success', 'Producto eliminado correctamente');
+        }
+    }
+    
+    private function isProductInFavorites($productId) {
+        // Verificar si hay registros en la tabla de favoritos que hagan referencia al producto
+        return Favorites::where('favorites.products_id', $productId)->exists();
+    }
